refactor(pre): add pdf prefix to preprocessors
