+++ src/pdf2aas/preprocessor/__init__.py
@@ -1,6 +1,6 @@
 from .core import Preprocessor
-from .pdf2htmlEX import PDF2HTMLEX, ReductionLevel
-from .pdfium import PDFium
+from .pdf_pdf2htmlEX import PDF2HTMLEX, ReductionLevel
+from .pdf_pdfium2 import PDFium
 from .text import Text
 
 __all__ = [
